Userspace memory dump support
* Implement userspace memory flattening

* Add example recipe for testing userspace memory flattening

* Bind to one CPU

* Add KFLAT test for userspace memory flattening

* Disable KASAN for kflat because it causes problems when dereferencing userspace pointers

---------

+++ tests/example_large_stringset.c
@@ -80,7 +80,7 @@ static int kflat_large_stringset_test(struct flat *flat) {
 	
 	unsigned i, j;
 	static const char chars[] = "ABCDEFGHIJKLMNOPQRST";
-	unsigned int r = 0;
+	unsigned long r = 0;
 
 	FLATTEN_SETUP_TEST(flat);
